- Added collectable objects deleting upon collection

+++ Coin_SMB3.cpp
@@ -21,8 +21,8 @@ Coin_SMB3::Coin_SMB3(const Vector2D spawnPosition
 , const char* const  filePathToSpriteSheet
 , const unsigned int amountOfSpritesOnWidth
 , const unsigned int amountOfSpritesOnHeight
-, const unsigned int collisionBoxWidth
-, const unsigned int collisionBoxHeight
+, const double       collisionBoxWidth
+, const double       collisionBoxHeight
 , const bool         collectableCanMove
 , const float        timePerAnimationFrame)
